Move get_data_analysis_type to analysis API
